requirements.txt
# Cecelia Semantic Brain - Dependencies

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Vector Database
chromadb>=0.4.22

# Embedding
openai>=1.10.0

# File Watching
watchdog>=4.0.0

# Text Processing
langchain-text-splitters>=0.0.1

# Core
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
